+++ tests/testproject/testapp/models.py
@@ -5,10 +5,28 @@ from __future__ import division, absolute_import, print_function, unicode_litera
 import os
 
 from django.db import models
-from save_the_change.mixins import SaveTheChange, TrackChanges
+from django.utils import six
+
+from save_the_change.mixins import SaveTheChange, TrackChanges, UpdateTogetherModel
 
 
-class EnlightenedModel(SaveTheChange, TrackChanges, models.Model):
+class Enlightenment(models.Model):
+	"""
+	A model to test ForeignKeys.
+	
+	"""
+	
+	aspect = models.CharField(max_length=32)
+
+
+class EnlightenedModel(SaveTheChange, TrackChanges, UpdateTogetherModel):
+	"""
+	A model to test (almost) everything else.
+	
+	TODO: Figure out a way to properly test {File,Image}Fields.
+	
+	"""
+	
 	big_integer = models.BigIntegerField()
 	boolean = models.BooleanField()
 	char = models.CharField(max_length=32)
@@ -17,6 +35,7 @@ class EnlightenedModel(SaveTheChange, TrackChanges, models.Model):
 	date_time = models.DateTimeField()
 	decimal = models.DecimalField(max_digits=16, decimal_places=8)
 	email = models.EmailField()
+	enlightenment = models.ForeignKey(Enlightenment)
 #	file = models.FileField(upload_to='./')
 	file_path = models.FilePathField(path=os.path.join(__file__, '..', 'uploads'))
 	float = models.FloatField()
@@ -32,3 +51,8 @@ class EnlightenedModel(SaveTheChange, TrackChanges, models.Model):
 	text = models.TextField()
 	time = models.TimeField()
 	URL = models.URLField()
+	
+	class Meta:
+		update_together = (
+			('big_integer', 'small_integer'),
+		)