+++ HISTORY.rst
@@ -1,6 +1,15 @@
 History
 =======
 
+1.1.0 (05/16/2014)
+------------------
+
+ - Add proper support for ForeignKeys (thanks to Brandon Konkle and
+   Brian Wilson).
+
+ - Add update_together field to model Meta, via UpdateTogetherModel.
+
+
 1.0.0 (09/08/2013)
 ------------------